+++ docs/whats_new.rst
@@ -26,6 +26,8 @@ Enhancements
 Bugs
 ~~~~
 - Fixing several models to be be compatibility with small windows size(:gh:`745` by `Bruno Aristimunha`_)
+- Fix initialization of :class:`braindecode.models.Deep4Net` (:gh:`752` by `Bruno Aristimunha`_)
+- Fix initialization of :class:`braindecode.models.ATCNet` (:gh:`754` by `Bruno Aristimunha`_)
 
 API changes
 ~~~~~~~~~~~